+++ R/schoenersD.R
@@ -1,10 +1,13 @@
 #' Calculates Schoener's D-value between two RasterLayers.
 #'
-#' RasterLayers must have identical resolutions and extents.
+#' RasterLayers must have identical resolutions and extents. The function will
+#' automatically normalize surfaces to sum to 1.
 #'
 #' Calculates similarity value of two RasterLayers using Schoener's D-metric.
 #' @param rast1 Input RasterLayer
 #' @param rast2 Input RasterLayer 2
+#'
+#' @importFrom raster cellStats
 #'
 #' @examples
 #' # Generate example probability surfaces.
@@ -36,9 +39,14 @@
 # Run pairwise comparisons.
 schoenersD <- function(rast1, rast2){
 
-  if(class(rast1) != "RasterLayer" | class(rast2) != "RasterLayer")
-    stop("arguments are not of class 'RasterLayer'")
+  stopifnot(
+    "Argument rast1 is not of class 'RasterLayer'" = class(rast1) == "RasterLayer" ,
+    "Argument rast2 is not of class 'RasterLayer'" = class(rast2) == "RasterLayer"
+  )
 
-  1 - (0.5 * raster::cellStats(abs(rast1 - rast2), stat='sum'))
+  if( cellStats(rast1, sum) != 1 ) { rast1 <- .findProductThenNormalize(rast1) }
+  if( cellStats(rast2, sum) != 1 ) { rast2 <- .findProductThenNormalize(rast2) }
+
+  1 - (0.5 * cellStats(abs(rast1 - rast2), stat='sum'))
 
   }