+++ NEWS.md
@@ -1,3 +1,10 @@
+# isocat 0.2.6.9000
+## New function
+* New function `makeMultiMonthIsoscape` to combine multiple monthly isoscapes and error maps into one of each, for making month-specific isoscape assignments.
+
+## Tweaks
+* `schoenersD` function now automatically normalizes surfaces to sum to 1, rather than just advising it in the documentation.
+
 # isocat 0.2.6
 
 ## Publication update